+++ slices/slice_test.go
@@ -18,22 +18,32 @@ var TSliceBenchs = []TSlice{
 	{
 		name: "10",
 		arg1: []interface{}{},
+		arg2: 5,
+		arg3: 10,
 	},
 	{
 		name: "100",
 		arg1: []interface{}{},
+		arg2: 50,
+		arg3: 100,
 	},
 	{
 		name: "1000",
 		arg1: []interface{}{},
+		arg2: 500,
+		arg3: 1000,
 	},
 	{
 		name: "10000",
 		arg1: []interface{}{},
+		arg2: 5000,
+		arg3: 10000,
 	},
 	{
 		name: "100000",
 		arg1: []interface{}{},
+		arg2: 50000,
+		arg3: 100000,
 	},
 }
 
@@ -43,7 +53,6 @@ func init() {
 		for j := 0; j < k/10; j++ {
 			TSliceBenchs[i].arg1 = append(TSliceBenchs[i].arg1, []interface{}{1, 2, 3, 4, 5, 6, 7, 8, 9, 0})
 		}
-
 	}
 }
 
@@ -61,21 +70,28 @@ func TestSlice(t *testing.T) {
 			arg1:     []interface{}{},
 			arg2:     2,
 			arg3:     6,
-			expected: []interface{}{},
+			expected: nil,
 		},
 		{
-			name:     "default",
+			name:     "normal",
 			arg1:     []interface{}{1, 2, 3, 4, 5, 6, 7, 8, 9, 0},
 			arg2:     2,
 			arg3:     5,
 			expected: []interface{}{3, 4, 5},
 		},
 		{
-			name:     "default1",
-			arg1:     []interface{}{"A", "B", "C", "D", "E", "F", "U"},
-			arg2:     1,
-			arg3:     6,
-			expected: []interface{}{"B", "C", "D", "E", "F"},
+			name:     "error-1",
+			arg1:     []interface{}{1, 2, 3, 4, 5, 6, 7, 8, 9, 0},
+			arg2:     2,
+			arg3:     100,
+			expected: nil,
+		},
+		{
+			name:     "error-2",
+			arg1:     []interface{}{1, 2, 3, 4, 5, 6, 7, 8, 9, 0},
+			arg2:     -5,
+			arg3:     3,
+			expected: nil,
 		},
 	}
 	for _, sample := range tests {
@@ -103,5 +119,4 @@ func BenchmarkSlice(b *testing.B) {
 			}
 		})
 	}
-
 }